==> notebooks/3-feature-engineering.ipynb <==
import numpy as np
import pandas as pd
from util import *

%load_ext autoreload

%reload_ext autoreload

## Run to generate train & test csv with features

train = pd.read_csv('../data/raw/train.csv')
test = pd.read_csv('../data/raw/test.csv')

train = process_basic(train)
test = process_basic(test)

train = process_flat_type(train)
test = process_flat_type(test)

comm = pd.read_csv("../data/processed/comm-dist-matrix.csv")
hawker = pd.read_csv("../data/processed/hawker-dist-matrix.csv")
mall = pd.read_csv("../data/processed/mall-dist-matrix.csv")
psch = pd.read_csv("../data/processed/psch-dist-matrix.csv")
ssch = pd.read_csv("../data/processed/ssch-dist-matrix.csv")
stations = pd.read_csv("../data/processed/stations-dist-matrix.csv")

test_comm = pd.read_csv("../data/processed/test-comm-dist-matrix.csv")
test_hawker = pd.read_csv("../data/processed/test-hawker-dist-matrix.csv")
test_mall = pd.read_csv("../data/processed/test-mall-dist-matrix.csv")
test_psch = pd.read_csv("../data/processed/test-psch-dist-matrix.csv")
test_ssch = pd.read_csv("../data/processed/test-ssch-dist-matrix.csv")
test_stations = pd.read_csv("../data/processed/test-stations-dist-matrix.csv")

## Point-of-interest integration

We take the distance to nearest POI instead of number of POI within

train['ssch_nearest_dist'] = ssch.drop(['Unnamed: 0'], axis=1).min(axis=1)
train['psch_nearest_dist'] = psch.drop(['Unnamed: 0'], axis=1).min(axis=1)
train['mall_nearest_dist'] = mall.drop(['Unnamed: 0'], axis=1).min(axis=1)
train['hawker_nearest_dist'] = hawker.drop(['Unnamed: 0'], axis=1).min(axis=1)
train['comm_nearest_dist'] = comm.drop(['Unnamed: 0'], axis=1).min(axis=1)
train['station_nearest_dist'] = stations.drop(['Unnamed: 0'], axis=1).min(axis=1)

test['ssch_nearest_dist'] = test_ssch.drop(['Unnamed: 0'], axis=1).min(axis=1)
test['psch_nearest_dist'] = test_psch.drop(['Unnamed: 0'], axis=1).min(axis=1)
test['mall_nearest_dist'] = test_mall.drop(['Unnamed: 0'], axis=1).min(axis=1)
test['hawker_nearest_dist'] = test_hawker.drop(['Unnamed: 0'], axis=1).min(axis=1)
test['comm_nearest_dist'] = test_comm.drop(['Unnamed: 0'], axis=1).min(axis=1)
test['station_nearest_dist'] = test_stations.drop(['Unnamed: 0'], axis=1).min(axis=1)

train = count_poi(train, ssch, 1, 'nbr_ssch_1k')
train = count_poi(train, psch, 1, 'nbr_psch_1k')
train = count_poi(train, mall, 1, 'nbr_mall_1k')
train = count_poi(train, hawker, 1, 'nbr_hawker_1k')
train = count_poi(train, comm, 1, 'nbr_comm_1k')
train = count_poi(train, stations, 1, 'nbr_station_1k')

test = count_poi(test, test_ssch, 1, 'nbr_ssch_1k')
test = count_poi(test, test_psch, 1, 'nbr_psch_1k')
test = count_poi(test, test_mall, 1, 'nbr_mall_1k')
test = count_poi(test, test_hawker, 1, 'nbr_hawker_1k')
test = count_poi(test, test_comm, 1, 'nbr_comm_1k')
test = count_poi(test, test_stations, 1, 'nbr_station_1k')

train = drop_cols(train)
test = drop_cols(test)

test

train

train.to_csv("../data/processed/train.csv")

test.to_csv("../data/processed/test.csv")
