==> notebooks/5-gridsearch.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from joblib import dump, load

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV, train_test_split

from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor

from sklearn.metrics import mean_squared_error

For comparing across ML models (i.e. RF, GB, LR), sample 20% and do 10-fold CV. Use default parameters for this.

For the top 2 models, use 60% of data to perform hyperparameter search.

For each model, we want the best set of parameters.

Output: Optimal parameters, as well as loss for each combination of parameters. Output: 10 different loss values/RMSE per model (boxplot)

train = pd.read_csv("../data/processed/train.csv")
print(train.shape)
train.drop(columns=['Unnamed: 0'], inplace=True)
train.head()

test = pd.read_csv("../data/processed/test.csv")
print(test.shape)
test.drop(columns=['Unnamed: 0'], inplace=True)
test.head()

y = train.resale_price
X = train.drop(columns=['resale_price'])
predictors = X.columns
print(X.shape)
print(y.shape)

seed = 42
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=seed)
print(X_train.shape)
print(y_train.shape)

print(X_test.shape)
print(y_test.shape)

seed = 42
X_train, _, y_train, _ = train_test_split(X_train, y_train, test_size=0.33, random_state=seed)
print(X_train.shape) # about 240k rows
print(y_train.shape)

print(X_test.shape)
print(y_test.shape)

sc = StandardScaler()
X_train = sc.fit_transform(X_train)
X_test = sc.transform(X_test)

%%time

grid = {
    'max_depth' : [1, 10, 100, 250, None],
    'n_estimators' : [1, 10, 100, 250]
}

est=RandomForestRegressor(random_state=seed)
print(est.get_params())

model = GridSearchCV(estimator=est,
                   param_grid=grid, scoring='neg_root_mean_squared_error',
                    verbose=3, n_jobs=-1)
model.fit(X_train, y_train)

# Store the parameters of the best model
best_params = model.best_params_

# Predict class labels of test data on the model with the best found parameters
y_pred = model.predict(X_test)

print(best_params)
print('Root Mean Squared Error RF:', np.sqrt(mean_squared_error(y_test, y_pred)))
dump(model.best_estimator_, 'rf.joblib')

### GBM CV
Follow https://www.analyticsvidhya.com/blog/2016/02/complete-guide-parameter-tuning-gradient-boosting-gbm-python/

Choose a relatively high learning rate. Generally the default value of 0.1 works but somewhere between 0.05 to 0.2 should work for different problems

Determine the optimum number of trees for this learning rate. This should range around 40-70. Remember to choose a value on which your system can work fairly fast. This is because it will be used for testing various scenarios and determining the tree parameters.

Tune tree-specific parameters for decided learning rate and number of trees. Note that we can choose different parameters to define a tree and I’ll take up an example here.

Lower the learning rate and increase the estimators proportionally to get more robust models.

%%time

grid = {
    'min_samples_split': [2, 4, 8, 16, 32],
    'max_depth': [1, 5, 10, 25, 50],
}

est=GradientBoostingRegressor(n_estimators = 50, random_state=seed)
print(est.get_params())

model2 = GridSearchCV(estimator=est,
                   param_grid=grid, scoring='neg_root_mean_squared_error',
                    verbose=3, n_jobs=-1)
model2.fit(X_train, y_train)

# Store the parameters of t3he best model
best_params2 = model2.best_params_

# Predict class labels of test data on the model with the best found parameters
y_pred2 = model2.predict(X_test)

model2.best_estimator_

print(best_params2)
print('Root Mean Squared Error GB:', np.sqrt(mean_squared_error(y_test, y_pred2)))

%%time

grid = {
    'learning_rate': [0.01, 0.001, 0.1, 0.05, 0.005],
    'n_estimators': [10, 25, 50, 100, 200],
}

est=GradientBoostingRegressor(max_depth=25, min_samples_split=32, random_state=seed)
print(est.get_params())

model2_2 = GridSearchCV(estimator=est,
                   param_grid=grid, scoring='neg_root_mean_squared_error',
                    verbose=3, n_jobs=-1)
model2_2.fit(X_train, y_train)

# Store the parameters of t3he best model
best_params2_2 = model2_2.best_params_

# Predict class labels of test data on the model with the best found parameters
y_pred2_2 = model2_2.predict(X_test)

print(best_params2_2)
print('Root Mean Squared Error GB:', np.sqrt(mean_squared_error(y_test, y_pred2_2)))
# dump(model2.best_estimator_, 'gbm.joblib')
