==> notebooks/6-train-predict.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from datetime import datetime
from joblib import dump, load

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV, train_test_split

from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from sklearn.linear_model import LinearRegression
from sklearn.experimental import enable_hist_gradient_boosting 
from sklearn.ensemble import HistGradientBoostingRegressor
from sklearn.metrics import mean_squared_error

seed = 42

train = pd.read_csv("../data/processed/train.csv")
print(train.shape)
train.head()

test = pd.read_csv("../data/processed/test.csv")
print(test.shape)
test.head()

y = train.resale_price
X = train.drop(columns=['resale_price'])
predictors = X.columns
print(X.shape)
print(y.shape)

sc = StandardScaler()
X = sc.fit_transform(X)
test = sc.transform(test)

# SPECIFY MODEL HERE
clf=GradientBoostingRegressor(learning_rate = 0.05, n_estimators= 200, max_depth=25, min_samples_split=32, random_state=seed)

clf.fit(X, y)
y_pred = clf.predict(test)

out = pd.DataFrame(data=y_pred).reset_index()
out.columns = ['Id', 'Predicted']
out.to_csv(f'../data/processed/submission_{datetime.now().strftime("%Y%m%d_%H%M%S")}.csv', index=False)

# dump(clf, 'gbm_19k.joblib')

out
